loding data

+++ src/app/home/branch/branch.component.ts
@@ -2,6 +2,8 @@ import { UserAuthentication } from 'src/app/shared/userAuthentication';
 import { BranchItem, HomeItem } from '../homeShared/homeItem';
 import { HomeDataService } from './../homeShared/home-data.service';
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { LoginService } from 'src/app/shared/login.service';
 
 @Component({
   selector: 'app-branch',
@@ -18,7 +20,9 @@ export class BranchComponent implements OnInit{
   id:number=0;
   deleteId:number=0;
   isShow:boolean=true;
-  constructor(private service:HomeDataService){
+  isUpdate:boolean=false;
+  // isPermission:any;
+  constructor(private service:HomeDataService,private curRoute:ActivatedRoute,private ls:LoginService){
     
   }
 
@@ -29,13 +33,23 @@ export class BranchComponent implements OnInit{
     this.data=JSON.parse(this.hello);
     console.log(this.data);
     this.isShow==true;
+    // this.isPermission= this.curRoute.snapshot.data['Permissions']
+    // console.log(this.isPermission)
+    // this.ls.routeData(this.isPermission);
   }
 
   deleteOpp(data:HomeItem){
     this.deleteId = this.service.branchDetail.findIndex((ele)=>{
       return ele.BranchId === data.BranchId && ele.BranchName === data.BranchName
       })
-    this.service.deleteFunBranch(this.deleteId);
+      if(this.isUpdate){
+        this.isShow=false;
+        this.service.deleteFunBranch(this.deleteId);
+      }
+      // else{
+        
+      //   this.service.deleteFunBranch(this.deleteId);
+      // }
   }
   
 
@@ -45,12 +59,18 @@ export class BranchComponent implements OnInit{
     })
     this.EditArray.push(val);
     this.isShow=true;
+
+    if(this.EditArray.length>1){
+      this.EditArray.shift();
+    }
+    this.isUpdate=true;
   }
 
   formEditData(BranchId:string,BranchName:string){
     this.service.editFunBranch(BranchName,BranchId,this.id);
     this.isShow=!this.isShow
     this.EditArray=[]
+    this.isUpdate=false;
   }
 
 }
