+++ src/app/home/company/company.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { HomeItem } from '../homeShared/homeItem';
 import { HomeDataService } from '../homeShared/home-data.service';
+import { LoginService } from 'src/app/shared/login.service';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-company',
@@ -15,20 +17,28 @@ export class CompanyComponent implements OnInit{
   id:number=0;
   deleteId:number=0;
   isShow:boolean=true;
-  constructor(private service:HomeDataService){}
+  isUpdate:boolean=false;
+  // isPermission:any;
+  constructor(private service:HomeDataService,private ls:LoginService,private curRoute:ActivatedRoute){}
 
   ngOnInit(){
     this.dataArray=this.service.companyDetail;
     this.hello=localStorage.getItem('keyPass')
     this.data=JSON.parse(this.hello);
     console.log(this.data);
+    // this.isPermission= this.curRoute.snapshot.data['Permissions']
+    // console.log(this.isPermission)
+    // this.ls.routeData(this.isPermission);
   }
 
   deleteOpp(data:HomeItem){
     this.deleteId = this.service.companyDetail.findIndex((ele)=>{
       return ele.CompanyId === data.CompanyId && ele.CompanyName === data.CompanyName
       })
-    this.service.deleteFunCompany(this.deleteId);
+      if(this.isUpdate){
+        this.isShow=false;
+        this.service.deleteFunCompany(this.deleteId);
+      }
   }
 
   onEditData(val:HomeItem){
@@ -37,11 +47,17 @@ export class CompanyComponent implements OnInit{
     })
     this.EditArray.push(val);
     this.isShow=true;
+
+    if(this.EditArray.length>1){
+      this.EditArray.shift();
+    }
+    this.isUpdate=true;
   }
 
   formEditData(CompanyId:string,CompanyName:string){
     this.service.editFunCompany(this.id,CompanyId,CompanyName);
     this.isShow=!this.isShow
     this.EditArray=[]
+    this.isUpdate=false;
   }
 }
